+++ Valet/MenuBar.swift
@@ -1,13 +1,13 @@
 import SwiftUI
 
 struct MenuBar: Commands {
-	@Binding var viewModes: ViewModes
+	@Binding var appState: AppState
 	var licenses: [License]
 
 	var body: some Commands {
 		CommandGroup(replacing: CommandGroupPlacement.newItem) {
 			Button("Add App", action: {
-				viewModes.showNewAppSheet.toggle()
+				appState.showNewAppSheet.toggle()
 			})
 			.keyboardShortcut(KeyboardShortcut(KeyEquivalent("N")))
 			Divider()
@@ -20,7 +20,7 @@ struct MenuBar: Commands {
 		}
 //		CommandGroup(replacing: CommandGroupPlacement.sidebar) {
 //			Button("Toggle Edit Mode", action: {
-//				viewModes.editMode.toggle()
+//				appState.editMode.toggle()
 //			})
 //			.keyboardShortcut(KeyboardShortcut(KeyEquivalent("E")))
 //			Divider()
